+++ routes/api/contacts.js
@@ -2,17 +2,7 @@ const express = require("express");
 const functions = require("../../models/contacts.js");
 const Joi = require("joi");
 const router = express.Router();
-const createError = (ERROR_TYPE, { message, data }) => {
-  return { type: ERROR_TYPE, message, data };
-};
-const ERROR_TYPES = {
-  BAD_REQUEST: "BAD_REQUEST",
-  UNAUTHORIZED: "UNAUTHORIZED",
-  FORBIDDEN: "FORBIDDEN",
-  NOT_FOUND: "NOT_FOUND",
-  INTERNAL_SERVER_ERROR: "INTERNAL_SERVER_ERROR",
-};
-
+const validateBody = require("../../validation.js");
 const standartBody = Joi.object({
   name: Joi.string(),
   email: Joi.string().email({
@@ -22,19 +12,6 @@ const standartBody = Joi.object({
   phone: Joi.number(),
   favorite: Joi.boolean(),
 });
-const validate = (target) => (schema) => (req, res, next) => {
-  const result = schema.validate(req[target]);
-  if (result.error) {
-    const custorError = createError(ERROR_TYPES.BAD_REQUEST, {
-      data: result.error,
-      message: result.error.message,
-    });
-    next(custorError);
-  } else {
-    next();
-  }
-};
-const validateBody = validate("body");
 router.get("/", async (req, res, next) => {
   try {
     const contacts = await functions.listContacts();
@@ -45,7 +22,6 @@ router.get("/", async (req, res, next) => {
     console.log(functions);
   }
 });
-
 router.get("/:contactId", async (req, res, next) => {
   const contact = await functions.getContactById(req.params.contactId);
   console.log(contact);
